Move config, lib, tasks into shared directory
