+++ .github/workflows/actions/install-node/action.yml
@@ -16,6 +16,7 @@ runs:
           app/node_modules
           docs/examples/*/node_modules
           package/node_modules
+          shared/*/node_modules
 
     - name: Setup Node.js
       uses: ./.github/workflows/actions/setup-node
