+++ docs/contributing/application-architecture.md
@@ -8,10 +8,6 @@
 
   Binary/executable files (i.e. bash scripts) mainly used in the [publishing process](/docs/releasing/publishing.md).
 
-- `config/`
-
-  Configuration files for the review app and [Jest](https://github.com/facebook/jest).
-
 - `dist/` **contains auto-generated files**
 
   Standalone builds of govuk-frontend. Provides a way to consume govuk-frontend without using npm.
@@ -20,19 +16,27 @@
 
   Documentation files.
 
-- `lib/`
-
-  Application modules and helpers.
-
 - `package/` **contains auto-generated files**
 
-  package published on npm.
+  Package published on npm.
   Consume all of govuk-frontend through a single package.
+
+- `shared/`
+
+  Shared packages used by tests, build tools and the [review app](../../app).
+
+  - `config/`
+
+    Configuration files for common paths, port numbers and [Jest](https://github.com/facebook/jest).
+
+  - `lib/`
+
+    Shared libraries for directory listings, component data, naming conventions.
+
+  - `tasks/`
+
+    See [tasks](tasks.md) for more information about the tasks.
 
 - `src/`
 
   Source files. See README.md in the src directory for details.
-
-- `tasks/`
-
-  Application modules and helpers. See [tasks](tasks.md) for more information about the tasks.
