+++ package.json.unit.test.js
@@ -1,5 +1,5 @@
 const packageLockJson = require('./package-lock.json')
-const packageJson = require('./tasks/package.json')
+const packageJson = require('./shared/tasks/package.json')
 
 describe('because rollup 0.60 drops support for Internet Explorer 8', () => {
   describe('rollup', () => {
