+++ shared/config/index.js
