+++ shared/config/jest/.eslintrc.js
