+++ shared/config/jest/browser/open.mjs
