+++ shared/config/jest/server/start.mjs
@@ -1,6 +1,6 @@
 import { setup } from 'jest-dev-server'
 
-import config from '../../../jest-dev-server.config.js'
+import config from '../../../../jest-dev-server.config.js'
 
 // Start web server(s)
 export default async () => {
