+++ shared/config/package.json
