+++ shared/lib/.eslintrc.js
