+++ shared/lib/helpers/files.js
