+++ shared/lib/helpers/names.js
