+++ shared/lib/helpers/puppeteer.js
