+++ shared/lib/index.js
