+++ shared/lib/package.json
