+++ shared/lib/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../tsconfig.base.json",
+  "extends": "../../tsconfig.base.json",
   "include": ["**/*.js", "**/*.mjs", "../config"],
   "compilerOptions": {
     "lib": ["ESNext", "DOM"]
