+++ shared/tasks/assets.mjs
