+++ shared/tasks/tsconfig.json
@@ -1,4 +1,4 @@
 {
-  "extends": "../tsconfig.base.json",
+  "extends": "../../tsconfig.base.json",
   "include": ["**/*.js", "**/*.mjs", "../config", "../lib"]
 }
