+++ src/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../tsconfig.base.json",
-  "include": ["**/*.js", "**/*.mjs", "../config", "../lib"],
+  "include": ["**/*.js", "**/*.mjs", "../shared/config", "../shared/lib"],
   "compilerOptions": {
     "lib": ["ESNext", "DOM"],
     "target": "ES2015",
